Styling & appearance fixes Added an ImageCropper preprocessor Added a PillowImage to annotations (original was changed to CVImage and both now inherit from an Abstract class named Image) - this allows easily dealing with animated pictures. ImageReader is agnostic to the image class type (it needs to be specified when it is initialized though) Added missing dependencies to requirements.txt, namely pillow and tf2onnx. TrainLogger console output is now supressed by default np.float --> float to support newer NumPy releases Updated .gitignore to exlude pyenv and idea settings

+++ mltu/dataProvider.py
@@ -11,6 +11,7 @@ from .transformers import Transformer
 import logging
 logging.basicConfig(format='%(asctime)s %(levelname)s %(name)s: %(message)s')
 
+
 class DataProvider:
     def __init__(
         self, 
@@ -238,7 +239,7 @@ class DataProvider:
 
         # Convert to numpy array if not already
         # TODO: This is a hack, need to fix this
-        if not isinstance(annotation, (np.ndarray, int, float, str, np.uint8, np.float)):
+        if not isinstance(annotation, (np.ndarray, int, float, str, np.uint8, float)):
             annotation = annotation.numpy()
 
         return data, annotation
